added min to logisticVRGDAPrices

+++ src/VRGDA/structs/LogisticProductParams.sol
@@ -8,8 +8,8 @@ import { LogisticVRGDAParams } from "./LogisticVRGDAParams.sol";
 /// @param decayConstant Precomputed constant that allows us to rewrite a pow() as an exp().
 /// @param pricingParams See `LogisticVRGDAParams`
 struct LogisticProductParams {
-  uint256 startTime;
-  uint256 startUnits;
-  int256 decayConstant;
+  uint40 startTime;
+  uint32 startUnits;
+  int184 decayConstant;
   mapping(address => LogisticVRGDAParams) pricingParams;
 }
